Major refactoring and preparing params for 3rd method (TTC)
